+++ tests/scripts/script_modifier_irc_in.py
@@ -1,0 +1,11 @@
+#
+# Author: Sébastien Helleu <flashcode@flashtux.org>
+#
+
+"""A WeeChat script."""
+
+import weechat
+
+if __name__ == '__main__':
+    if weechat.register('script', 'author', '0.1', 'GPL3', 'desc', '', ''):
+        weechat.hook_modifier('irc_in_privmsg', 'callback', '')