+++ ChangeLog.md
@@ -4,8 +4,9 @@
 
 - Sort messages by severity (high first).
 - Fix duplicate errors on `python2_bin`.
-- Add warnings on deprecated info `irc_nick_color` and `irc_nick_color_name`.
 - Check also empty files.
+- Add warning on modifier `irc_in_xxx`.
+- Add warnings on deprecated info `irc_nick_color` and `irc_nick_color_name`.
 - Add warnings on deprecated functions `hook_completion_get_string` and `hook_completion_list_add`.
 
 ## Version 0.1.0 (2021-04-19)